New Class: Message And Updated Documentation
Documentation:
This commit updates the documentation to show the index.html file which contains an example and some information about this project.

New Class: Message
This commit adds a class Message which is a large improvement but also is required to repeat for each method we have. This has been done for better 'flow' and also allows us to use jsdocs better as it is used for classes.