Merge Dev into Stable (#17)
* Turned embeds off, errors before processing.

* Added comment note

* New Embed Class

New Embed Class with plenty of documentation and some fixes to TextChannel#Send
Some attempts to obtain roles too

* quick fixz

* Added permissions class and put onto ./index.js

* Revert "Added permissions class and put onto ./index.js"

This reverts commit 0fbbaf980785ae9478b62b9f1d996ae6de74cca7.

* Added permissions, put into ./index.js

* Fexed my dude

* I'll remove the whiteline too

* Some fixes

Starting to make some methods return the classes

* Documentation for Permissions Class + New Method

Added some documentation for Permissions Class and also added a new method to turn an array of Permissions to the bitfield.  Now we need  to get a way to convert a bitfield to array of permissions.