Rather than our custom function to emit events, we now use node.js to do
This update optimizes node.js' "events" module to allow us to emit events with only one callback/arg. This also merges the Connect.js file to the Client.js file which means less files and also allows people to have more documentation and also to be more optimized for a user endpoint as it could also allow them to understand the source code better. This also brings client.users as it didn't work before as it had an empty collection but now it should have all of the CACHED users.